requirements.txt
google-genai
pydantic
json_repair
fix_busted_json

